requirements.txt
discord
asyncio
pymongo
time
json
dnspython
dataclasses
python-dotenv